rpm단위 우선 float으로 변경

+++ calc_cadence_speed/calc_cadence_speed.ino
@@ -11,7 +11,7 @@ int lcd_update_before = 0;
 int lcd_update_interval = 300; // 300ms마다 업데이트
 int rpm_update_before[2] = {0, 0};
 int rpm_update_interval = 1000; //300ms마다 업데이트
-int rpm_arr[2] = {0, 0};
+float rpm_arr[2] = {0.0, 0.0};
 
 void rpm_fun1() {// 인터럽트로 카운트 증가하는 함수
   rpmcount[0]++;
@@ -24,7 +24,7 @@ void rpm_fun2() {
 void (*rpm_fun[2])() = {rpm_fun1, rpm_fun2};    // int형 반환값, int형 매개변수 두 개가 있는 함수 포인터 배열 선언, 첫 번째 요소에 함수의 메모리 주소 저장
 void rpm_calc(int i);
 void lcd_init();
-void lcd_update(int rpm[]);
+void lcd_update(float rpm[]);
 
 void setup() {
   Serial.begin(115200);
@@ -62,7 +62,7 @@ void lcd_update(int rpm[]) {
     lcd.setCursor(5, 2);
     lcd.print("   "); // LCD에 새로운 rpm값을 계속 표시해주기 위해서 이전에 표시되었던 내용을 지워주는 과정
     lcd.setCursor(5, 2);   // 5,2에서 글쓰기 시작
-    lcd.print(rpm[0]);
+    lcd.print((int)rpm[0]);
     Serial.println(rpm[0]);
     lcd.setCursor(9, 2);
     lcd.print("rpm");      // lcd 모니터에 rpm 표시해주기
@@ -70,8 +70,8 @@ void lcd_update(int rpm[]) {
     lcd.setCursor(5, 3);
     lcd.print("   ");    //LCD에 새로운 rpm값을 계속 표시해주기 위해서 이전에 표시되었던 내용을 지워주는 과정
     lcd.setCursor(5, 3);   // 5,3에서 글쓰기 시작
-    int bikespeed = int( (rpm[1] * circumference * 60 ) / (1000 * 1000));
-    lcd.print(bikespeed);
+    float bikespeed = (rpm[1] * circumference * 60 ) / (1000 * 1000);
+    lcd.print((int)bikespeed);
     Serial.println(bikespeed);
     lcd.setCursor(9, 3);
     lcd.print("km/h");
@@ -82,7 +82,7 @@ void rpm_calc(int i) {
   if ( (millis() - rpm_update_before[i]) >= rpm_update_interval ) {
     rpm_update_before[i] = millis();
     detachInterrupt(i);
-    rpm_arr[i] = int( 60000 * (rpmcount[i] / hall_num[i]) / (millis() - timeold[i]) );
+    rpm_arr[i] = 60000 * (rpmcount[i] / hall_num[i]) / (millis() - timeold[i]);
     timeold[i] = millis(); // 기준시간 리셋
     rpmcount[i] = 0;
     attachInterrupt(i, rpm_fun[i], FALLING); // 인터럽트 0->2번핀 1->3번핀
